text file update

+++ include/Evaluare.h
@@ -10,17 +10,17 @@ class Evaluare
     std::string tip;
     float parte_din_final;
     float punctaj_maxim;
-    int prag;
+    float prag;
     float nota;
 
 public:
-    Evaluare(const std::string, const float, const float, const int, const float);
+    Evaluare(const std::string, const float, const float, const float, const float);
 
     std::string getTip() const;
     float getParteFinal() const;
     float getMaxim() const;
     float getNota() const;
-    int getPrag() const;
+    float getPrag() const;
 
     void setNota(const float);
 };
